backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
motor>=4.0.0
pydantic>=2.3.0
groq>=0.1.0
